+++ src/main/java/com/visitorproject/repo/UserAdressesRepo.java
@@ -3,5 +3,5 @@ package com.visitorproject.repo;
 import com.visitorproject.entity.UserAddresses;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserAdressesRepo extends JpaRepository<UserAddresses, Integer> {
+public interface UserAdressesRepo extends JpaRepository<UserAddresses, Long> {
 }
